accuracy

+++ lucid/lucid/renderer/vector/rect.h
@@ -1,10 +1,10 @@
 #pragma once
 
 struct rect_t {
-	float x, y, w, h;
+	double x, y, w, h;
 
 	rect_t() {}
-	rect_t(float x, float y, float w, float h) {
+	rect_t(double x, double y, double w, double h) {
 		this->x = x;
 		this->y = y;
 		this->w = w;
